libmachine: move copy file function where it is used, in the driver
