+++ pkg/libmachine/mcnutils/utils.go
@@ -2,37 +2,8 @@ package mcnutils
 
 import (
 	"fmt"
-	"io"
-	"os"
 	"time"
 )
-
-func CopyFile(src, dst string) error {
-	in, err := os.Open(src)
-	if err != nil {
-		return err
-	}
-
-	defer in.Close()
-
-	out, err := os.Create(dst)
-	if err != nil {
-		return err
-	}
-
-	defer out.Close()
-
-	if _, err = io.Copy(out, in); err != nil {
-		return err
-	}
-
-	fi, err := os.Stat(src)
-	if err != nil {
-		return err
-	}
-
-	return os.Chmod(dst, fi.Mode())
-}
 
 func WaitForSpecificOrError(f func() (bool, error), maxAttempts int, waitInterval time.Duration) error {
 	for i := 0; i < maxAttempts; i++ {
